api/requirements.txt
absl-py==0.9.0
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
aiosqlite==0.12.0
asn1crypto==1.3.0
astor==0.8.0
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
beautifulsoup4==4.9.0
blinker==1.4
cachetools==3.1.1
cchardet==2.1.6
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
cycler==0.10.0
databases==0.3.1
elasticsearch==7.6.0
fake-useragent==0.1.11
fastapi==0.54.1
fire==0.3.0
gast==0.2.2
geographiclib==1.50
geopy==1.21.0
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googletransx==2.4.2
grpcio==1.27.2
h11==0.9.0
h5py==2.10.0
idna==2.9
kiwisolver==1.2.0
Markdown==3.1.1
matplotlib==3.2.1
multidict==4.7.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.3
pg8000==1.15.2
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycares==3.1.1
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-multipart==0.0.5
pytz==2020.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
schedule==0.6.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
starlette==0.13.2
termcolor==1.1.0
toposort==1.5
tqdm==4.45.0
tweepy==3.8.0
twint==2.1.20
urllib3==1.25.8
uvicorn==0.11.4
websockets==8.1
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yarl==1.4.2
